requirements.txt
Flask==1.1.2
matplotlib==3.5.2
mnist==0.2.2
numpy==1.22.3
Pillow==9.4.0
scipy==1.8.1
torchvision==0.14.1
